+++ Tests/Functional/SpeakingSettingsTest.php
@@ -37,7 +37,7 @@ use Bithost\Pdfviewhelpers\Exception\ValidationException;
  *
  * @author Markus Mächler <markus.maechler@bithost.ch>, Esteban Gehring <esteban.gehring@bithost.ch>
  */
-class SpeakingSettingsTest extends AbstractFunctionalTest
+class SpeakingSettingsTest extends AbstractFunctionalTestCase
 {
     protected array $typoScriptFiles = [
         'EXT:pdfviewhelpers/Tests/Functional/Fixtures/SpeakingSettings/setup.typoscript',