+++ Tests/Functional/ViewHelpers/AvoidPageBreakInsideCustomFontViewHelperTest.php
@@ -30,14 +30,14 @@ namespace Bithost\Pdfviewhelpers\Tests\Functional\ViewHelpers;
  *  This copyright notice MUST APPEAR in all copies of the script!
  * * */
 
-use Bithost\Pdfviewhelpers\Tests\Functional\AbstractFunctionalTest;
+use Bithost\Pdfviewhelpers\Tests\Functional\AbstractFunctionalTestCase;
 
 /**
  * AvoidPageBreakInsideCustomFontViewHelperTest
  *
  * @author Markus Mächler <markus.maechler@bithost.ch>, Esteban Gehring <esteban.gehring@bithost.ch>
  */
-class AvoidPageBreakInsideCustomFontViewHelperTest extends AbstractFunctionalTest
+class AvoidPageBreakInsideCustomFontViewHelperTest extends AbstractFunctionalTestCase
 {
     protected array $typoScriptFiles = [
         'EXT:pdfviewhelpers/Tests/Functional/Fixtures/AvoidPageBreakInsideViewHelper/CustomFont.typoscript',