+++ ext_emconf.php
@@ -15,7 +15,7 @@ $EM_CONF[$_EXTKEY] = [
 	'version' => '3.0.0',
 	'constraints' => [
 		'depends' => [
-			'typo3' => '10.4.0-11.5.99',
+			'typo3' => '11.5.0-12.4.99',
 			'php' => '7.4.0-0.0.0',
 		],
 		'conflicts' => [],